Add unitTest for checking sibling relationship
A unit test has been added to databaseTest to see if any siblings in a tree shuold actually be parents. For now, I have only implemented it for thermo groups. It should be implemented to kinetics, solvation, statmech and transport as well, but I don't think we have any commits of RMG-database that will pass those tests.